Update to C++ 20 Fix integer casts Remove file dialog

+++ sdlgui/common.cpp
@@ -16,12 +16,12 @@
 #if defined(_WIN32)
 #include <SDL.h>
 #else
-#include <SDL2/SDL.h>
+#include <SDL.h>
 #endif
 #if defined(_WIN32)
 #include <SDL_image.h>
 #else
-#include <SDL2/SDL_image.h>
+#include <SDL_image.h>
 #endif
 #include "nanovg.h"
 
@@ -87,7 +87,11 @@ PntRect clip_rects(PntRect a, const PntRect& b)
 
 SDL_Color Color::toSdlColor() const
 {
-  SDL_Color color{ r() * 255, g() * 255, b() * 255, a() * 255 };
+  SDL_Color color{
+      static_cast<Uint8>(r() * 255),
+      static_cast<Uint8>(g() * 255),
+      static_cast<Uint8>(b() * 255),
+      static_cast<Uint8>(a() * 255) };
   return color;
 }
